src/main/requirements.txt
###### AUTO-GENERATED Requirements file for: AskAI ######

hspylib>=1.12.40
hspylib-clitt>=0.9.126
hspylib-setman>=0.10.35
langchain>=0.2.7
langchain-openai>=0.1.14
langchain-community>=0.2.7
langchain-google-community>=1.0.6
langchainhub>=0.1.20
faiss-cpu~=1.8.0
openai-whisper>=20231117
openai>=1.35.12
google-api-python-client>=2.137.0
fake_useragent>=1.5.1
rich>=13.7.1
textual>=0.73.0
soundfile>=0.12.1
PyAudio>=0.2.14
SpeechRecognition>=3.10.4
opencv-python>=4.10.0.84
argostranslate==1.9.1
chromadb>=0.5.3
retry2>=0.9.5
pause>=0.3
requests>=2.32.3
urllib3>=1.26.19
protobuf>=4.22.1
torch>=2.2.0
tqdm>=4.66.4
torchvision>=0.17.0
transformers>=4.42.4
stanza>=1.1.1
unstructured>=0.14.10
unstructured[md]>=0.14.10
python-magic-bin>=0.4.14
tiktoken>=0.7.0
aiohttp>=3.9.3
html2text>=2024.2.26
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-proto>=1.22.0
pytz>=2024.1
bs4>=0.0.2
